requirements-test.txt
# Test dependencies - includes main requirements needed for testing
# This file ensures test automation has all required dependencies

# Main requirements (core dependencies needed for tests)
-r backend/requirements.txt

# Additional testing tools
pytest-mock>=3.11.0
pytest-docker>=3.2.0
docker>=7.1.0
mutmut>=2.4.0
hypothesis>=6.0.0

# Missing test dependencies (Issue #984)
PyJWT>=2.8.0
psutil>=5.9.0

# Machine Learning dependencies (needed for test collection)
numpy>=1.21.0
scipy>=1.8.0
scikit-learn>=1.3.0
torch>=2.0.0
joblib>=1.3.0
aiofiles>=23.0.0

# Development tools
pre-commit>=3.0.0